+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "connect-nodejs-samples",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Samples of using Spotware Connect Open API in JavaScript and Node.js",
   "keywords": ["spotware", "connect", "ctratder"],
   "author": "Spotware Services <connect@spotware.com>",
@@ -9,6 +9,6 @@
     "protobufjs": "^4.0.0",
     "byte": "^1.1.4",
     "chalk": "^1.1.1",
-    "protobuf-stream":"spotware/node-protobuf-stream#master"
+    "protobuf-stream":"spotware/node-protobuf-stream"
   }
 }
